+++ src/models/message.go
@@ -8,7 +8,7 @@ type Message struct {
 	UserID          int            `json:"user_id" gorm:"not null"`
 	CommodityItem   *CommodityItem `gorm:"constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
 	CommodityItemID int            `json:"commodity_item_id"` //如果不constraint join时要注意
-	CurrentPrice    float64        `json:"current_price" gorm:"not null"`
+	CurrentPrice    float32        `json:"current_price" gorm:"not null"`
 	CreateAt        time.Time      `json:"create_at"`
 	//PriceLimit   float64
 }
